\name{RDSTK-package}
\alias{RDSTK-package}
\alias{RDSTK}
\docType{package}
\title{RDSTK: A R wrapper for the Data Science Toolkit API}
\description{
This package contains several functions that provide direct access to the Data Science Toolkit API.  See www.datasciencetoolkit.org for an overview of the API.  The package is an attempt to R-ify calls to this API.}

\details{
\tabular{ll}{
Package: \tab RDSTK\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-30\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}
}

\author{
Ryan Elmore
Maintainer: Ryan Elmore <rtelmore@gmail.com>
}

\references{
http://www.datasciencetoolkit.org
}

\examples{
ip2coordinates("134.184.34.17, 48.82.68.161")
}
