% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-to-sentences.R
\name{text2sentences}
\alias{text2sentences}
\title{Identifies sentences in a text string.}
\usage{
text2sentences(text, session = RCurl::getCurlHandle())
}
\arguments{
\item{text}{A string (hopefully) containing sentences.}

\item{session}{The CURLHandle object giving the structure for the options 
and that will process the command. For curlMultiPerform, this is an object 
of class code MultiCURLHandle-class.}
}
\value{
A list containing
    \item{sentences}{A string identifying the sentences in the text.}
}
\description{
This function returns the legitimate sentences (if they exist) from a text
 string.
}
\examples{
\dontrun{
sentences <- "But this does, it contains enough words. So does this 
              one, it appears correct. This is long and complete enough
              too."
text2sentences(sentences)
}

}
\references{
http://www.datasciencetoolkit.org/developerdocs#text2sentences
}
\seealso{
\code{\link{curlPerform}},
  \code{\link{getCurlHandle}},
  \code{\link{dynCurlReader}}
}
