% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates-to-politics.R
\name{coordinates2politics}
\alias{coordinates2politics}
\title{Coverts latitude and longitude coordinates to politics expressions.}
\usage{
coordinates2politics(latitude, longitude, session = RCurl::getCurlHandle())
}
\arguments{
\item{latitude}{The latitude (numeric) of the point you wish to reference.}

\item{longitude}{The longitude (numeric) of the point you wish to reference.}

\item{session}{The CURLHandle object giving the structure for the options 
and that will process the command. For curlMultiPerform, this is an object 
of class code MultiCURLHandle-class.}
}
\value{
Returns a JSON string.
}
\description{
A function to return the countries, states, provinces, cities, 
  constituencies and neighborhoods that the latitude and longitude 
  point lies within
}
\examples{
\dontrun{
coordinates2politics(37.769456, -122.429128)
}

}
\references{
http://www.datasciencetoolkit.org/developerdocs#coordinates2politics
}
\seealso{
\code{\link{getURL}}, \code{\link{getCurlHandle}}
}
