% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDSTK-pkg.R
\docType{package}
\name{RDSTK}
\alias{RDSTK}
\alias{RDSTK-package}
\title{RDSTK: R Interface for the Data Science Toolkit}
\description{
This package contains several functions that provide direct access to the 
Data Science Toolkit API.  See www.datasciencetoolkit.org for an overview 
of the API.  The package is an attempt to R-ify calls to this API.

By default the packages accesses the API at www.datasciencetoolkit.org. 
Alternatively, because it is possible to clone the DSTK service on a local 
machine, you can point the package to an alternate API using 
\code{options("RDSTK_api_base"="http://localhost:8080")}.
}
\section{RDSTK Functions}{

  \code{\link{coordinates2politics}},
  \code{\link{coordinates2statistics}},
  \code{\link{html2text}},
  \code{\link{ip2coordinates}},
  \code{\link{street2coordinates}},
  \code{\link{text2people}},
  \code{\link{text2sentences}},
  \code{\link{text2sentiment}},
  \code{\link{text2times}}
}

